+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ## Unreleased
 
+### Features
+
+- Add `Sentry.setAttributes` and `Sentry.removeAttribute` ([#3352](https://github.com/getsentry/sentry-dart/pull/3352))
+  - These attributes are set at the scope level and apply to all logs (and later to metrics and spans).
+  - When a scope attribute conflicts with a log-level attribute, the log-level attribute always takes precedence.
+
 ### Fixes
 
 - Android app crashing on hot-restart in debug mode ([#3358](https://github.com/getsentry/sentry-dart/pull/3358))